# Optimize byte<->array conversions in to_ndarray