Optimize the byte->nparray conversion by using `np.frombytes` instead of
`struct.unpack`. Unfortunately, we still have to make a copy because `bytes`
aren't mutable so we have to create a `bytearray` copy. There may be some way
to eliminate this with `memoryview`, but I'm not familiar with that yet.

Then optimize pdarray->byte conversations by directly copying the localized
array's buffer into a bytes type instead of writing to and then reading from a
memory mapped file. Here again there's an extra copy, but we could eliminate
that by using the `makeArrayFromPtr` interface.

Note that this breaks portability when the endianness of the server doesn't
match the client. Previously, everything forced big-endian and now we assume
client and server match. This can probably be addressed by sending the
endianness along with the message and interpreting differently, but I'm not
sure. Because of this breakage I've drafted this for now.

Here are some performance results for to_ndarray on 16-node-xc

| config        | Time  |  Rate     |
| ------------- | ----: | --------: |
| before        | 39.0s |  20 MiB/s |
| np.frombuffer | 23.8s |  32 MiB/s |
| direct-bytes  |  2.1s | 362 MiB/s |