@ronawho 2021-06-30
@reuster986 this is a draft PR that pretty significantly improves `to_ndarray` performance more as part of https://github.com/Bears-R-Us/arkouda/issues/794. It breaks portability when the endianness of the server doesn't match the client, which is why I opened as a draft. Previously, everything forced big-endian, but now it just relies on native endianness.

I think we can do something like send the endianness as part of the message and convert if needed, but I'm not positive yet. I wanted to check if this is worth pursuing (or if things seem fast enough without this) and how much you care about things working when the client/server have different endianness. So far as I know the only practical case where that would matter is if you're running on a powerpc server and have your client running on an x86 laptop or something. I'm not sure how important that is, but it seems like a shame to break the functionality anyways.