@reuster986 2021-07-01
@ronawho Cool! I think this is definitely worth pursuing, as there are a lot of workflows that pass data back and forth between arkouda and numpy/pandas. The 15x performance improvement is too good to pass up, and I think the potential memory savings you mentioned are also quite attractive. As server-client transfers become faster, users will probably start doing bigger and bigger transfers, which will start to put pressure on the memory of the head node and the client.

I don't know of anyone using arkouda where the client and server have different endianness, but it's bound to happen at some point, so it's good to plan for it now. Of course, it's difficult to find a platform to test on... I'm wondering if the best thing to do for now is 

1. Have `ServerConfig.chpl` store the endianness of the server so that it can be retrieved by the client with `ak.get_config()`
2. The client can detect its own endianness with `sys.byteorder`
3. If the two don't match, have the client fall back to using `struct.unpack` and plugging in the endianness of the server.

This would delegate the endianness handling to `struct`, at the expense of some performance, but the mixed-endian case is so rare that I think it's ok.