@ronawho 2021-07-01
Sounds good. I have access to a powerpc system so I should be able to test the mixed endianness. Do you remember if there was a specific reason for changing to using big-endian in https://github.com/Bears-R-Us/arkouda/commit/a375b5933674ed777bb7d98b1aa4a861c67c0911? I suspect not, but want to make sure I'm not missing something that would make this approach obviously wrong.

For now I'll go with the plan you outlined and in followup PRs optimize to `ak.array()` case and look to minimize the number of copies that are needed. Do you have a sense for whether `ak.array()` or `to_nparray()` is more common or are the used about the same?