@ronawho 2021-07-02
Ah, so it turns out the powerpc system I have access to is little-endian (and most ppc after power 7 are little-endian.) I can still test this by using the current file writing trick on the server, but I think this makes it very unlikely anybody will have mixed endian client/server in practice.