@ronawho 2021-07-02
I'm going to close this in favor of multiple PRs so we can better track perf changes in nightly configs. I expect there to be PRs for:
 - Optimizing client conversations between bytes and numpy arrays 
 - Optimizing server conversations between bytes and pdarrays 
 - Reducing the number of memory copies needed in the client